calcNutrition/requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
pyocr==0.8.3
